requirements.txt
pysqlite3-binary~=0.5.4
python-fasthtml~=0.10.0
bcrypt~=4.2.1